requirements.txt
numpy
Pillow
torch
plotly
scikit-image
torchvision
matplotlib
